% Generated by build script
\name{unhcr_demographics}
\alias{unhcr_demographics}
\title{Fetch UNHCR demographic data}
\usage{
unhcr_demographics()
}
\arguments{
\item{year}{Year (e.g. 2023). Required.}
\item{country_origin}{ISO3 country of origin code}
\item{country_asylum}{ISO3 country of asylum code}
\item{limit}{Max results (default 100)}
}
\value{tibble: year, country_origin, country_asylum, sex, age_group, value}
\description{Fetch UNHCR demographic data}
